requirements.txt
Pillow==3.2.0
backports-abc==0.4
bcrypt==2.0.0
beautifulsoup4==4.4.1
cffi==1.7.0
chardet==2.2.1
colorama==0.2.5
greenlet==0.4.10
html5lib==0.999
lxml==3.6.0
motor==0.6.2
pycparser==2.14
pymongo==2.8
requests==2.2.1
six==1.5.2
tornado==4.3
urllib3==1.7.1
wheel==0.24.0
